requirements.txt
# Backend dependencies
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
pydantic<2.0
SQLAlchemy<2.0
alembic==1.13.2
python-dotenv==1.0.1
aiosqlite==0.17.0

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7

# Frontend (documented here, installed with npm/yarn, not pip)
# React 18+, Vite, TailwindCSS, shadcn/ui, Recharts or Plotly
